Add ProfileObservable for use in new TV target (#1595)
In order to highlight how coupled the UI targets are to Partout, here we
start restoring the explicit imports. The pain point is the
ObservableObject implementations, because they both do business logic
and manage UI state. Take the opportunity to migrate to the Observation
framework (bump targets accordingly).

Therefore, for each ObservableObject:

- Move view actions/state out to a new Observable class that doesn't
depend on Partout
- Only use local domain entities in the public Observable interfaces,
even if the wrappers are as naive as AppProfile
- Replace ObservableObject classes with Observable counterparts in views

Eventually, this process will make all the views agnostic of Partout.
Don't replace the views directly, rather duplicate them for dynamic
switch with a config flag.

This PR:

- Aggregates more domain entities into CommonLibrary
- Adds the AppProfile[Header] domain entity for public interfaces
- Creates ProfileObservable out of ProfileManager
- Duplicates the AppLibraryTV target
- Replaces all occurrences of ProfileManager with ProfileObservable in
AppLibraryTV

App behavior is not affected at this stage. Only changes in active code:

- ProfileManager
  - Simplify allRemoteProfiles to be only remoteProfileIds
- Rename profile() to partoutProfile() (deprecated, use AppProfile
later)
  - Add import() method (unused in legacy views)

Deprecate all legacy code, esp. prepare to delete RegistryCoder now that
Registry is inside ProfileManager.