Group domain entities of future ABI (#1601)
The ABI namespace will involve the entities that the views handle to
communicate with the business logic.

Therefore:

- Move library-only entities to the "Internal" folder (will make
internal later)
  - AppRelease
  - ConfigBundle
  - ProfileAttributes
- Move view-only entities back to AppLibrary
  - AppTip
  - EditableProfile
  - HTTPAddressPort
  - OnboardingStep
  - ProfilesLayout
  - SystemAppearance

Anything else is enclosed in the ABI enum namespace in the "Domain"
folder. These will be the alphabet of the ABI layer.

This PR is safe as it only moves files around.