Use AppConfiguration from SwiftUI environment (#1604)
Formerly Constants and BundleConfiguration, hide them behind a
statically-initialized value type AppConfiguration, also injected into
the SwiftUI environment for use in the views. Catch bundle
misconfiguration early, and unify access to global app constants. The
Dependencies struct also ceases to be a singleton, it is instantiated in
AppContext (app) and PacketTunnelProvider (tunnel).