Build non-Apple targets with CMake (#1616)
After splitting app-apple and app-shared properly in #1612, the time has
come to draft a concrete starting point for the Windows/Linux app, with
app-android already being set up for Android. In order to mix Swift with
other languages and dependencies, and in a cross-platform fashion, the
preferred build system is **CMake**. SwiftPM is not convenient yet for
such a complex build.

## Business logic

The cross-platform business logic is written in Swift, and lies fully in
the **app-shared** folder. This library layer internally depends on
Partout and will provide:

- A direct Swift ABI to Apple frontends (SwiftUI)
- C/JNI bindings to the ABI for non-Swift frontends (Jetpack Compose and
wxWidgets)

## Apps

There are three frontend targets:

- **app-apple**: (Xcode + SwiftPM) SwiftUI app built with a few
Apple-specific targets, then app-shared, and Partout. Third parties
(OpenSSL, WireGuard, etc.) come from Partout as XCFrameworks.
- **app-android**: (Android Studio + CMake) Depends on Swift up to
app-shared (ABI), Jetpack Compose from the ABI up. The ABI is deployed
to the app as a dynamic library, then Kotlin takes over at the upper
layers.
- **app-cross**: (CMake) Like app-android, it depends on app-shared. The
app is built together with the ABI and therefore statically linked. The
UI is implemented with wxWidgets, which additionally depends on GTK+ on
Linux. Windows requires no extra dependencies.

## Distribution

The CMake project generates the output to `bin/<platform-arch>`, where
third parties are distributed as dynamic libraries. The Windows/Linux
apps are also generated in the `app` subfolder, with CMake also bundling
the local Swift runtime. Wintun is also bundled on Windows.

Ideally, the `bin/<platform-arch>/app` output folder includes all the
required libraries to run the app on a clean Linux/Windows installation.
This was only tested on arm64 machines.

## Gotchas

Not sure about some of these:

- Android: Reuse the CMake toolchain and swiftly wrapper from Partout.
Make sure to define the required environment variables in `env.sh`.
- Windows: Use VC 2022 generator for wxWidgets, Ninja is problematic.
Force release to avoid the "d" suffix.
- Windows: Do not strip the executable, it will break.
- Windows: Distribute OpenSSL DLLs from the "bin" folder, not "lib".
- Windows: Link to absolute library paths to cope with the inconsitent
"lib" prefix.
- Linux: Add `$ORIGIN` to the runpath to ensure proper library lookup
without LD_LIBRARY_PATH.
- macOS (Development): Importing Partout as a Swift module in app-shared
requires the modulemaps of all C modules in the headers search paths.
- Not sure at all about wxWidgets compile/linking flags, but they will
be clearer along the way.