Review cross-platform app and ABI distribution (#1632)
First off, force link to bundled libraries by specifying absolute paths.
`libpassepartout` was linking to generic `ssl` and `crypto`, which are
often available system-wide, and as a result, the Android app was
randomly crashing on OpenSSL calls from the OpenVPN code.

The likely cause was that the library was linked to some system-wide
OpenSSL at build time, but the app was eventually calling functions of
the bundled OpenSSL libraries. The binary mismatch would lead to random
crashes.

After that:

- Distribute and strip binaries consistently, also the dynamic
libraries, while retaining the unstripped versions
- Simplify and clean up the CommonLibrary ABI targets
- Design the initial Passepartout ABI calls in passepartout.h
- Migrate all the VPN-related JNI code from the Android app to Partout,
so that the app only interacts with the ABI