Migrate AppContext logic to ABI (#1634)
Interpose an ABI layer between AppContext (app-apple) and the business
logic of the managers (app-shared).

Currently, the managers embed the UI state, and the views use/observe
them directly. This architecture is not ideal for sharing the most code
with the non-Apple apps, because it's hard to track how the views
interact with the business objects. Instead, everything should happen
through a common ABI interface (ABIProtocol), Swift on Apple and Swift/C
elsewhere.

Therefore, this flow:

```
AppContext (observable managers + events logic) → Managers
```

Is changed to:

```
AppContext (observables) → ABI (events logic) → Managers
```

So that the views will eventually communicate with the observables and
be fully agnostic of the underlying ABI and the managers. The UI state
of the managers remains deprecated and will be removed after migrating
SwiftUI to use the new observables.

This PR moves the ownership of all the managers from AppContext to the
ABI:

- AppContext creates observables from the managers, plus the view-only
onboarding observable
- AppContext registers the general events callback with the ABI
- The ABI becomes the single event observer of all the managers
- Observables no longer observe manager events actively
- Instead, events are forwarded to AppContext, which then dispatches
them to the proper observables onUpdate()
- Launch assertions are moved to
AppContext.assertMissingImplementations()
- All the app lifecycle handlers, like onApplicationActive() and other
event handlers, are moved to CommonABI, of which AppContext is now a
thin wrapper for SwiftUI

Necessary decoupling:

- Move sorting comparators from ProviderRepository to AppLibrary,
because KeyPathComparator comes from Foundation
- Minimal changes to include CommonProviders in cross-platform too
(exclude scripting API impl and tests)
- Decouple APIManager and PreferencesManager from ObservableObject
- SystemExtensionManager in ABI as generic ExtensionInstaller
- Move manager events to ABI event namespace
- Make the C ABI part of Core
- Spit the PSP_CROSS into PSP_MONOLITH (build with CMake) and PSP_CROSS
(build on non-Apple, even with SwiftPM)
- Migrate all tests to Swift Testing to decouple from Foundation
(implied by XCTest)