Move observables logic to ABI singleton (#1639)
Most of the observables, namely:

- AppEncoder
- Config
- IAP
- Profile
- Tunnel
- VersionChecker
- WebReceiver

No longer access managers directly. The logic is transferred to
CommonABI, so that the observables invoke ABI methods directly. This
will help translate the business methods to a C ABI, which the native UI
will call into from C++ (wx) and Kotlin/JNI (Compose). The active UI
must not be affected by this PR because it doesn't use the observables,
except for AppEncoderObservable in a few places.

Other fixes in app-cross:

- psp_init() crashing due to wrong arguments
- Wrong destination of passepartout.h