Transfer app objects initialization to ABI (#1645)
Move the objects initialization code from AppContext+Production to a
convenience initializer in AppABI+Apple with just two arguments:

- ABI.AppConfiguration
- KeyValueStore

The app configuration and the key-value store are enough to build all
the necessary app objects. All the code from Dependencies is moved to
the AppConfiguration+Dependencies extensions.

Other than that:

- Move minor logic and formatting out of ABI
- Make ConfigManager.activeFlags sync without MainActor hack
- Rename sysexManager to extensionInstaller
- Create dummy IAPManager and APIManager
- Do not force-unwrap C args in ABI
- Fix a bug in encoding of experimental preferences

The goal of this PR is to ease the transition to psp_app_init() for
cross-platform, in that it highlights the most needed input to create
the app objects for the AppABI: the app configuration JSON and a
persistent key-value store.